[#56] Moved appcompat classes out of core
