+++ core-appcompat/build.gradle
@@ -1,0 +1,38 @@
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+    id 'appyx-publish-android'
+}
+
+android {
+    namespace = 'com.bumble.appyx.core.appcompat'
+    compileSdk 32
+
+    defaultConfig {
+        minSdk 21
+        targetSdk 32
+
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_1_8
+    }
+    lintOptions {
+        abortOnError false
+    }
+    testOptions {
+        unitTests.all {
+            // interface method default implementation
+            exclude '**/*$DefaultImpls.class'
+        }
+    }
+}
+
+dependencies {
+    api project(':core')
+    api libs.androidx.appcompat
+}
