+++ core/build.gradle
@@ -48,8 +48,6 @@ dependencies {
     api libs.compose.runtime
 
     implementation libs.androidx.activity.compose
-    implementation libs.androidx.core
-    implementation libs.androidx.appcompat
     implementation libs.androidx.lifecycle.java8
     implementation libs.androidx.lifecycle.runtime
     implementation libs.compose.foundation.layout
