+++ core/src/main/java/com/bumble/appyx/core/integrationpoint/permissionrequester/PermissionRequesterHost.kt
@@ -4,7 +4,6 @@ import android.app.Activity
 import android.content.pm.PackageManager
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
 
 interface PermissionRequesterHost {
 
@@ -24,20 +23,6 @@ interface PermissionRequesterHost {
 
         override fun requestPermissions(requestCode: Int, permissions: Array<String>) {
             ActivityCompat.requestPermissions(activity, permissions, requestCode)
-        }
-
-    }
-
-    class FragmentHost(private val fragment: Fragment) : PermissionRequesterHost {
-
-        override fun isGranted(permission: String): Boolean =
-            ContextCompat.checkSelfPermission(fragment.requireContext(), permission) == PackageManager.PERMISSION_GRANTED
-
-        override fun shouldShowRationale(permission: String): Boolean =
-            fragment.shouldShowRequestPermissionRationale(permission)
-
-        override fun requestPermissions(requestCode: Int, permissions: Array<String>) {
-            fragment.requestPermissions(permissions, requestCode)
         }
 
     }
